+++ ErsatzTV.Core/Domain/MediaSource/PlexMediaSource.cs
@@ -11,4 +11,5 @@ public class PlexMediaSource : MediaSource
     // public bool IsOwned { get; set; }
     public List<PlexConnection> Connections { get; set; }
     public List<PlexPathReplacement> PathReplacements { get; set; }
+    public DateTime? LastCollectionsScan { get; set; }
 }