+++ ErsatzTV.Infrastructure/Data/Configurations/Collection/PlexCollectionConfiguration.cs
@@ -1,0 +1,10 @@
+﻿using ErsatzTV.Core.Domain;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace ErsatzTV.Infrastructure.Data.Configurations;
+
+public class PlexCollectionConfiguration : IEntityTypeConfiguration<PlexCollection>
+{
+    public void Configure(EntityTypeBuilder<PlexCollection> builder) => builder.ToTable("PlexCollection");
+}