+++ ErsatzTV.Infrastructure/Plex/PlexEtag.cs
@@ -253,6 +253,31 @@ public class PlexEtag
         byte[] hash = SHA1.Create().ComputeHash(ms);
         return BitConverter.ToString(hash).Replace("-", string.Empty);
     }
+    
+    public string ForCollection(PlexMetadataResponse response)
+    {
+        using MemoryStream ms = _recyclableMemoryStreamManager.GetStream();
+        using var bw = new BinaryWriter(ms);
+
+        // collection key
+        bw.Write(response.Key);
+
+        // collection added at
+        bw.Write(response.AddedAt);
+
+        // collection updated at
+        bw.Write(response.UpdatedAt);
+        
+        // collection child count
+        bw.Write(response.ChildCount ?? "0");
+
+        // collection is smart collection
+        bw.Write(response.Smart ?? "0");
+
+        ms.Position = 0;
+        byte[] hash = SHA1.Create().ComputeHash(ms);
+        return BitConverter.ToString(hash).Replace("-", string.Empty);
+    }
 
     private enum FieldKey : byte
     {
@@ -270,6 +295,9 @@ public class PlexEtag
         Thumb = 20,
         Art = 21,
 
-        File = 30
+        File = 30,
+        
+        ChildCount = 40,
+        Smart = 41 // smart collection bool
     }
 }